Creation of the option to view the Clusters obtained from the publications of the researchers. Files created: cloudcluster.html, cloudCluster.js, cloudCluster.css and datos_clustering.json (to get data offline).
Updates in the files app.js, controllers.js and services.js to support this new option. All the controllers and imports required by AngularJS are created in those files.

Changes in the html to include the links to the new option in the following pages: index, cloudgroup, genericcloud, home, map-sources, search, tags.