Changes to update small bugs in the webapp code that appeared after the parametrization of languages and the update of endpoints (in which the datatypes where added to each field).
Updates to improve visualization. 

Now the view "Clusters by keyword" does not return to the "Home" page while processing the request for the clusters that contain a keyword.

Titles were added to the "keyword cloud" and "Clusters by keyword" pages.

Translation of additional labels, and some titles that are defined in a Sparql were translated into Spanish. The translation files were completed.

Changes that allow the user to click a different language and the system does not go back to the Home page. Now the application remains in the same web page, only changing the language.