Updates to improve visualization in the WebApp module
Changes implemented in the option "Group by keywords" that allow the user to click over each node that represents a publication, and the application shows the information about that publications as well as the authors that are related to it, both by clustering and by publication. 

Improvements in the translation.