Updates in the WebApp module
Keyword Cloud: Adding a new filter by Research Area, and when the user selects a Research Area, the application shows a cloud with other related areas (according to similarities). Make the fonts of the cloud more visible and drawing bigger circles.

Adding export to Excel: Export the information of the Author and his/her publications.

Adding export options of some information in RDF, N3, JSON, Turtle.

Performance Improvements: disable debug options in the webapp

Changes in the Cluster by Keywords: Make the fonts of the cloud more visible and drawing bigger circles.