Updates in the web app:
Author's Search Tree: Added the option to export list of publication by author to PDF.

Adding the User's Manual to the Help Page. It can be shown in PDF and downloaded.

More changes in the images and text of the application to reflect the new name of the plataform and other updates.

Adding some labels to the web page.

Correction of the problem that was showing the word "undefined" at the beginning of the field "Keywords" in the list of authors of the option "Clusters by Keyword". Now the word "undefined" is not shown anymore.

Translation of a few more tags in the application.