Implementing the classes for translation of text in the platform:
In order to translate texts inside our platform, we implemented some classes and functions that allow us to use the Microsoft Translator API, which is free if we do not exceed certain limit (unlike Google Translator API). To use this translation service, we just need to call the function from any part of the platform.