Add view to see related authors and solve problem of scrolling.
In many cases an author has many related authors; then, the view gets overwhealmed of text, and it is difficult to visualize. For this reason, the first 20 relavant related authors are shown, and the remaining ones are shown in the view.

The view was scrolling down when loading. Now this problem has been solved.

Signed-off-by: Xavier Sumba <xavier.sumba93@ucuenca.ec>